LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/canvas/offscreen/manual/filter/offscreencanvas.filter-expected.txt


FAIL none assert_array_equals: The image data generated by filter blur(10px) should be the same for both OffscreenCanvas and regular canvas expected property 0 to be 0 but got 108 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "108,39,29,26,106,35,26,29,112,40,32,32,112,45,30,34,110,48,28,37")
